Annotate the Rust idioms for use as a teaching example
Mark the two startup expects as the binary's only deliberate
panics and state the rule at the site: after the first request
is accepted, failures become responses or log lines, never
panics.

Comment the idioms where they live: lazy unwrap_or_else
defaults, thread ownership of the Connection instead of a
Mutex, the blocking_recv/await bridge between sync and async,
let _ = for deliberate discards, collecting Result rows into
Result<Vec<_>>, degrade-not-panic on corrupt detail blobs,
u32 over a guarded u64 cast, serde(default), and escaping
client strings at the markup boundary.

+++ src/model.rs
@@ -9,7 +9,13 @@ use serde::{Deserialize, Serialize};
 pub struct NewRun {
     pub kind: String,
     pub host: String,
-    pub duration_ms: u64,
+    /// `u32` on purpose: ~49 days of milliseconds is more than any
+    /// run needs, and the conversion into `SQLite`'s `i64` is
+    /// lossless. Prefer tightening a type over guarding a cast —
+    /// `i64::from(u32)` cannot fail, so nothing needs checking.
+    pub duration_ms: u32,
+    /// `#[serde(default)]` makes the field optional on the wire; an
+    /// absent `detail` arrives as `Value::Null` instead of a 400.
     #[serde(default)]
     pub detail: serde_json::Value,
 }
